api/character-user-link extract EnumField from the mod_guild.forms to common.forms
